# contracts/admin-api/v2/openapi.json
{
  "components": {
    "schemas": {
      "AdminCapability": {
        "enum": [
          "admin_read_v2",
          "admin_device_manage_v2",
          "admin_wechat_status_v2",
          "admin_wechat_manage_v2",
          "admin_wechat_login_v2",
          "admin_maintenance_v2",
          "admin_results_manage_v2"
        ],
        "type": "string"
      },
...
      "AdminErrorV2": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "type": "object"
      },
...
      "AdminMetaV2": {
        "additionalProperties": false,
        "properties": {
          "adminApiVersion": {
            "example": 2,
            "format": "int32",
            "maximum": 2,
            "minimum": 2,
            "type": "integer"
          },
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/AdminCapability"
            },
            "type": "array"
          },
          "generatedAt": {
            "type": "integer"
          },
          "relayVersion": {
            "type": "string"
          },
          "schemaVersion": {
            "example": 2,
            "format": "int32",
            "maximum": 2,
            "minimum": 2,
            "type": "integer"
          }
        },
        "required": [
          "schemaVersion",
          "generatedAt",
          "adminApiVersion",
          "relayVersion",
          "capabilities"
        ],
        "type": "object"
      },
...
      "EmptyMutationRequestV2": {
        "additionalProperties": false,
        "type": "object"
      },
...
      "ResultDetailV2": {
        "additionalProperties": false,
        "properties": {
          "acceptedAt": {
            "type": "integer"
          },
          "bodyBytes": {
            "minimum": 0,
            "type": "integer"
          },
          "bodyPurgedAt": {
            "type": [
              "integer",
              "null"
            ]
          },
          "completedAt": {
            "type": [
              "integer",
              "null"
            ]
          },
          "contentMode": {
            "type": "string"
          },
          "correlationKey": {
            "type": "string"
          },
          "durationMs": {
            "type": [
              "integer",
              "null"
            ]
          },
          "notificationAttemptCount": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "notificationId": {
            "type": "string"
          },
          "notificationLastErrorCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "notificationStatus": {
            "$ref": "#/components/schemas/ResultNotificationStatusV2"
          },
          "ownerDeviceId": {
            "type": "string"
          },
          "pageExpiresAt": {
            "type": "integer"
          },
          "pageState": {
            "$ref": "#/components/schemas/ResultPageStateV2"
          },
          "resultId": {
            "type": "string"
          },
          "resultRevision": {
            "type": "string"
          },
          "resultRowId": {
            "description": "Relay's globally unique results-row identifier. Admin commands use this\nkey because result_id is scoped to its owning device.",
            "type": "string"
          },
          "revokedAt": {
            "type": [
              "integer",
              "null"
            ]
          },
          "safeTitle": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "sourceHash": {
            "type": "string"
          },
          "startedAt": {
            "type": [
              "integer",
              "null"
            ]
          },
          "targetAccountFingerprint": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedAt": {
            "type": "integer"
          }
        },
        "required": [
          "resultRowId",
          "resultId",
          "ownerDeviceId",
          "safeTitle",
          "source",
          "contentMode",
          "sourceHash",
          "acceptedAt",
          "updatedAt",
          "pageState",
          "pageExpiresAt",
          "bodyBytes",
          "notificationId",
          "notificationStatus",
          "notificationAttemptCount",
          "correlationKey",
          "resultRevision"
        ],
        "type": "object"
      },
      "ResultListItemV2": {
        "additionalProperties": false,
        "properties": {
          "acceptedAt": {
            "type": "integer"
          },
          "bodyBytes": {
            "minimum": 0,
            "type": "integer"
          },
          "bodyPurgedAt": {
            "type": [
              "integer",
              "null"
            ]
          },
          "contentMode": {
            "type": "string"
          },
          "notificationAttemptCount": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "notificationId": {
            "type": "string"
          },
          "notificationLastErrorCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "notificationStatus": {
            "$ref": "#/components/schemas/ResultNotificationStatusV2"
          },
          "ownerDeviceId": {
            "type": "string"
          },
          "pageExpiresAt": {
            "type": "integer"
          },
          "pageState": {
            "$ref": "#/components/schemas/ResultPageStateV2"
          },
          "resultId": {
            "type": "string"
          },
          "resultRowId": {
            "description": "Relay's globally unique results-row identifier. Admin commands use this\nkey because result_id is scoped to its owning device.",
            "type": "string"
          },
          "safeTitle": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "sourceHash": {
            "type": "string"
          },
          "targetAccountFingerprint": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedAt": {
            "type": "integer"
          }
        },
        "required": [
          "resultRowId",
          "resultId",
          "ownerDeviceId",
          "safeTitle",
          "source",
          "contentMode",
          "sourceHash",
          "acceptedAt",
          "updatedAt",
          "pageState",
          "pageExpiresAt",
          "bodyBytes",
          "notificationId",
          "notificationStatus",
          "notificationAttemptCount"
        ],
        "type": "object"
      },
      "ResultNotificationStatusV2": {
        "enum": [
          "pending_channel",
          "sending_channel",
          "retry_wait",
          "provider_accepted",
          "blocked_activation",
          "blocked_reconnect",
          "dead_letter",
          "expired",
          "cancelled"
        ],
        "type": "string"
      },
      "ResultPageStateV2": {
        "description": "A result-page projection deliberately excludes plaintext body, encrypted\nstorage, share bearer tokens and resolved URLs. Those values never belong\non an Admin list or detail response.",
        "enum": [
          "available",
          "revoked",
          "expired",
          "content_unavailable"
        ],
        "type": "string"
      },
...
      "ResultsPageV2": {
        "additionalProperties": false,
        "properties": {
          "generatedAt": {
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResultListItemV2"
            },
            "type": "array"
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "total": {
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "generatedAt"
        ],
        "type": "object"
      },
...
      "WechatTestReceiptV2": {
        "additionalProperties": false,
        "properties": {
          "issuedAt": {
            "type": "integer"
          },
          "receiptId": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/WechatTestStateV2"
          }
        },
        "required": [
          "receiptId",
          "state",
          "issuedAt"
        ],
        "type": "object"
      },
      "WechatTestStateV2": {
        "enum": [
          "accepted_by_relay"
        ],
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Restricted same-origin Relay operator API. Rust wire types are the only editable contract authority.",
    "license": {
      "name": ""
    },
    "title": "PromptDock Relay Admin API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/api/v2/meta": {
      "get": {
        "operationId": "getAdminMetaV2",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMetaV2"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorV2"
                }
              }
            },
            "description": "Admin API error"
          }
        },
        "tags": [
          "admin-v2"
        ]
      }
    },
...
    "/admin/api/v2/results": {
      "get": {
        "operationId": "listAdminResultsV2",
        "parameters": [
          {
            "in": "query",
            "name": "pageState",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ResultPageStateV2"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultsPageV2"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorV2"
                }
              }
            },
            "description": "Admin API error"
          }
        },
        "tags": [
          "admin-v2"
        ]
      }
    },
    "/admin/api/v2/results/{result_row_id}": {
      "get": {
        "operationId": "getAdminResultV2",
        "parameters": [
          {
            "in": "path",
            "name": "result_row_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultDetailV2"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorV2"
                }
              }
            },
            "description": "Admin API error"
          }
        },
        "tags": [
          "admin-v2"
        ]
      }
    },
...
    "/admin/api/v2/wechat/test": {
      "post": {
        "operationId": "testAdminWechatV2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyMutationRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WechatTestReceiptV2"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorV2"
                }
              }
            },
            "description": "Admin API error"
          }
        },
        "tags": [
          "admin-v2"
        ]
      }
    }
  },
...
}